+++ packages/expo-image-manipulator/ios/ImageManipulatorUtils.swift
@@ -1,4 +1,4 @@
-import SDWebImageWebPCoder
+internal import SDWebImageWebPCoder
 import Photos
 import ExpoModulesCore