+++ packages/expo/CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ### 💡 Others
 
+- [iOS] Updated AppDelegate Swift header imports for xcframework compatibility. ([#44248](https://github.com/expo/expo/pull/44248) by [@chrfalch](https://github.com/chrfalch))
+- [iOS] Remove `RCTHostRuntimeDelegate` usage now that it's merged into `RCTHostDelegate`. ([#43838](https://github.com/expo/expo/pull/43838) by [@zoontek](https://github.com/zoontek))
+
 ## 55.0.9 — 2026-03-27
 
 ### 🎉 New features